report 增加对话列表，详情区分对话类型
