+++ apps/web/src/components/mf/ChatLayout/index.tsx
@@ -13,29 +13,13 @@ import React, {
 import { clsx } from 'clsx'
 import { useRouter } from 'next/router'
 import { useWorkspaces } from '@/hooks/useWorkspaces'
-import { NoData } from '@/components/mf/NoData'
 
 import styles from './index.module.scss'
-import ScrollBar from '@/components/ScrollBar'
 
-import Logo from '@/icons/mind-flow.svg'
 import { useSession } from '@/hooks/useAuth'
-import { Popover, PopoverButton, PopoverPanel } from '@headlessui/react'
-import {
-  AlertDialog,
-  AlertDialogAction,
-  AlertDialogCancel,
-  AlertDialogContent,
-  AlertDialogDescription,
-  AlertDialogFooter,
-  AlertDialogHeader,
-  AlertDialogTitle,
-  AlertDialogTrigger,
-} from '@/components/Dialog'
-import { useDeleteChat } from '@/hooks/mf/chat/useChatDelete'
+
 import { showToast } from '../Toast'
-import { useChatEdit } from '@/hooks/mf/chat/useChatEdit'
-import Spin from '@/components/Spin'
+
 import ArrowRight from '@/icons/arrow-right-line.svg'
 import { useChatRoundCreate } from '@/hooks/mf/chat/useChatSessionCreate'
 import { v4 as uuidv4 } from 'uuid'
@@ -43,50 +27,8 @@ import { ChatType, MessageContent, useChatDetail } from '@/hooks/mf/chat/useChat
 import { useChatStop } from '@/hooks/mf/chat/useChatStop'
 import { ChatStatus } from '@/hooks/mf/chat/useChatStatus'
 import { useChatCreate } from '@/hooks/mf/chat/useCreateChat'
-
+import ChatListBox from '../ChatList'
 const defaultMsg: MessageContent = { id: '', role: 'system', content: '我是你的AI小助手' }
-interface Item {
-  type: 'del' | 'edit'
-  label: string
-}
-
-interface IMoreBtnProps {
-  items: Item[]
-  onItemClick?: (type: 'del' | 'edit') => void
-}
-
-const MoreBtn = ({ items, onItemClick }: IMoreBtnProps) => {
-  const [isOpen, setIsOpen] = useState(false)
-
-  return (
-    <Popover className={styles.moreOpt}>
-      <PopoverButton as="div" onClick={() => setIsOpen(!isOpen)}>
-        <img src="/icons/more.svg" width={16} />
-      </PopoverButton>
-      <PopoverPanel
-        anchor="bottom"
-        className={clsx('shadow-lg', styles.morePopoverLayout)}
-        style={{ marginTop: '8px' }}>
-        {({ close }) => (
-          <div className={styles.moreBtnLayout}>
-            {items.map((item, index) => (
-              <div
-                className={styles.moreBtn}
-                key={index}
-                onClick={(e) => {
-                  e.stopPropagation()
-                  close()
-                  onItemClick && onItemClick(item.type)
-                }}>
-                {item.label}
-              </div>
-            ))}
-          </div>
-        )}
-      </PopoverPanel>
-    </Popover>
-  )
-}
 
 interface Props {
   children: React.ReactNode
@@ -403,34 +345,14 @@ export const useChatLayoutContext = () => {
 }
 
 export default function ChatLayout({ children }: Props) {
-  const [updateTitleEvent, setUpdateTitleEvent] = useState<EventSource | null>(null)
-  const [isDialogOpen, setDialogOpen] = useState(false)
-
   const [routeTitle, setRouteTitle] = useState('')
-  const [currentTitle, setCurrentTitle] = useState('')
-  const [isCommit, setIsCommit] = useState(false)
-  const delChatId = useRef('')
-  const eventTimeoutId = useRef(-1)
-  const lastedTimeoutId = useRef(-1)
-  const lastedEvent = useRef<EventSource | null>(updateTitleEvent)
+
   const [workspaces] = useWorkspaces()
   const session = useSession()
   const router = useRouter()
-  const [{ deleteChat }] = useDeleteChat()
-  const [{ editTitle }] = useChatEdit()
-  const { chatList, setChatList, setRoundList, refreshChatList } = useChatLayoutContext()
 
   const firstLetter = session.data?.loginName.charAt(0).toUpperCase() // 获取用户名的第一个字母并转为大写
   const chatId = router.query.chatId
-  useEffect(() => {
-    refreshChatList()
-    titleUpdate()
-    return () => {
-      if (lastedEvent.current) {
-        lastedEvent.current.close()
-      }
-    }
-  }, [])
 
   useEffect(() => {
     setRouteTitle(() => {
@@ -442,187 +364,14 @@ export default function ChatLayout({ children }: Props) {
     })
   }, [router])
 
-  useEffect(() => {
-    lastedTimeoutId.current = eventTimeoutId.current
-  }, [eventTimeoutId])
-
-  useEffect(() => {
-    lastedEvent.current = updateTitleEvent
-  }, [updateTitleEvent])
-
-  const titleUpdate = useCallback(() => {
-    const eventSource = new EventSource(
-      `${process.env.NEXT_PUBLIC_API_URL}/v1/mf/chat/title/update`,
-      {
-        withCredentials: true, // 如果需要发送 cookies
-      }
-    )
-    eventSource.onmessage = (event: MessageEvent) => {
-      try {
-        const data = JSON.parse(event.data)
-        setChatList((prevList) => {
-          const lastIndex = prevList.findIndex((item) => item.id === data.chatId) // 获取最后一条消息的索引
-          const updatedList = [...prevList]
-          if (lastIndex >= 0) {
-            updatedList[lastIndex] = {
-              ...updatedList[lastIndex],
-              title: (updatedList[lastIndex].title = data.title), // 将新内容追加到最后一条消息
-            }
-          }
-          return updatedList
-        })
-      } catch (e) {}
-    }
-    eventSource.onerror = (error) => {
-      eventSource.close()
-      eventTimeoutId.current = window.setTimeout(() => {
-        titleUpdate()
-      }, 5000)
-    }
-    setUpdateTitleEvent(eventSource)
-  }, [chatList])
-
-  const deleteChatById = (id: string) => {
-    deleteChat(id).then(() => {
-      setDialogOpen(false)
-      setChatList((prevChatList) => prevChatList.filter((chat) => chat.id !== id))
-      showToast('删除成功', 'success')
-      if (chatId === id) {
-        router.replace('/home')
-      }
-    })
-  }
-
-  const commitTitle = (id: string, title: string) => {
-    setIsCommit(true)
-    return editTitle(id, title)
-      .then(() => {
-        showToast('对话更新成功', 'success')
-      })
-      .finally(() => {
-        setCurrentTitle('')
-        setIsCommit(false)
-      })
-  }
-
-  const updateChat = (chat: HistoryChat) => {
-    if (chat.title !== currentTitle) {
-      commitTitle(chat.id, currentTitle)
-        .then(() => {
-          chat.title = currentTitle
-        })
-        .finally(() => {
-          chat.isEditing = false
-          setChatList((prevItems) =>
-            prevItems.map((item) => (item.id === chat.id ? { ...item, newChat: chat } : item))
-          )
-        })
-    } else {
-      chat.isEditing = false
-      setChatList((prevItems) =>
-        prevItems.map((item) => (item.id === chat.id ? { ...item, newChat: chat } : item))
-      )
-    }
-  }
-
-  const ChatListItem = ({ chat }: { chat: HistoryChat }) => {
-    return (
-      <div
-        className={clsx(
-          styles.chatItem,
-          chat.id === chatId ? styles.active : '' // 添加选中样式
-        )}
-        onClick={() => {
-          if (chat.id === chatId || chat.isEditing) {
-            return
-          }
-          setRoundList([])
-
-          if (chat.type === 'report') {
-            router.push(
-              `/workspaces/${workspaces.data[0].id}/documents/${chat.documentId}/notebook/edit?chatId=${chat.id}`
-            )
-          } else {
-            router.push(`/rag/${chat.id}`)
-          }
-        }}>
-        {chat.isEditing ? (
-          <div className={styles.inputBox}>
-            <input
-              type="text"
-              className={styles.titleInput}
-              onChange={(e) => {
-                setCurrentTitle(e.target.value)
-              }}
-              onKeyDown={(e: React.KeyboardEvent<HTMLInputElement>) => {
-                if (e.key === 'Enter' && !e.shiftKey) {
-                  e.preventDefault() // 防止默认的换行行为
-                  updateChat(chat)
-                }
-              }}
-              value={currentTitle}
-              onBlur={() => {
-                updateChat(chat)
-              }}
-              autoFocus
-            />
-            <div className={isCommit ? styles.loadingIcon : styles.loadingIconHidden}>
-              <Spin color="#2F69FE" wrapperClassName="pl-2" />
-            </div>
-          </div>
-        ) : (
-          <div className={styles.itemTitle}>{chat.title}</div>
-        )}
-        <MoreBtn
-          items={[
-            { type: 'edit', label: '编辑标题' },
-            { type: 'del', label: '删除' },
-          ]}
-          onItemClick={(type) => {
-            if (type === 'del') {
-              setDialogOpen(true)
-              delChatId.current = chat.id
-            } else if (type === 'edit') {
-              setCurrentTitle(chat.title)
-              setChatList((prevItems) =>
-                prevItems.map((item) => (item.id === chat.id ? { ...item, isEditing: true } : item))
-              )
-            }
-          }}
-        />
-      </div>
-    )
+  if (!workspaces || workspaces.data.length === 0) {
+    return
   }
 
   return (
     <div className={clsx(styles.chatLayout)}>
       <div className={clsx(styles.left, 'text-sm')}>
-        <div className={styles.top}>
-          <div
-            className={clsx('flex w-full flex-col', styles.logo_icon)}
-            onClick={() => {
-              router.push('/home')
-            }}>
-            <Logo />
-          </div>
-          <div
-            className={styles.createBtn}
-            onClick={() => {
-              router.push(`/home`)
-            }}>
-            <img src="/icons/chat-new-line.svg" width={16} height={16} />
-            <span>新建对话</span>
-          </div>
-        </div>
-        <ScrollBar className={styles.chatListWrapper}>
-          {chatList.length ? (
-            chatList.map((chat) => {
-              return <ChatListItem key={chat.id} chat={chat} />
-            })
-          ) : (
-            <NoData className={styles.empty} />
-          )}
-        </ScrollBar>
+        <ChatListBox chatId={chatId?.toString() || ''} workspaceId={workspaces.data[0].id} />
       </div>
       <div className={styles.main}>
         <div className={styles.mainTop}>
@@ -650,25 +399,6 @@ export default function ChatLayout({ children }: Props) {
         </div>
         {children}
       </div>
-      <AlertDialog open={isDialogOpen}>
-        <AlertDialogContent>
-          <AlertDialogHeader>
-            <AlertDialogTitle>删除对话</AlertDialogTitle>
-            <AlertDialogDescription>确定删除该对话么？</AlertDialogDescription>
-          </AlertDialogHeader>
-          <AlertDialogFooter>
-            <AlertDialogCancel>取消</AlertDialogCancel>
-            <AlertDialogAction
-              onClick={() => {
-                if (delChatId.current) {
-                  deleteChatById(delChatId.current)
-                }
-              }}>
-              确定
-            </AlertDialogAction>
-          </AlertDialogFooter>
-        </AlertDialogContent>
-      </AlertDialog>
     </div>
   )
 }
