+++ apps/web/src/components/mf/ChatList/index.module.scss
@@ -1,0 +1,104 @@
+.chatList {
+  background-color: #f4f7fe;
+  height: 100%;
+  overflow: hidden;
+  display: flex;
+  flex-direction: column;
+  flex-shrink: 0;
+  width: 15rem;
+
+  .top {
+    padding: 16px;
+    flex-shrink: 0;
+
+    .createBtn {
+      margin-top: 1.25rem;
+      display: flex;
+      width: 13rem;
+      height: 2.5rem;
+      border-radius: 0.5rem;
+      align-items: center;
+      padding: 0.625rem 0.75rem;
+      gap: 0.5rem;
+      color: #2f69fe;
+      background: linear-gradient(0deg, rgba(47, 105, 254, 0.12), rgba(47, 105, 254, 0.12)), #ffffff;
+      border: 1px solid rgba(47, 105, 254, 0.12);
+      cursor: pointer;
+    }
+  }
+
+  .chatListWrapper {
+    height: 100%;
+    padding: 0 16px 16px 16px;
+    flex-grow: 1;
+    overflow: auto;
+
+    .chatItem {
+      position: relative;
+      display: flex;
+      flex-direction: row;
+      align-items: center;
+      border-radius: 0.5rem;
+      padding: 0.75rem 2.5rem 0.75rem 0.75rem;
+      cursor: pointer;
+
+      .itemTitle {
+        text-overflow: ellipsis;
+        overflow: hidden;
+        white-space: nowrap;
+        flex-grow: 1;
+      }
+      .inputBox {
+        display: flex;
+        justify-content: center;
+        align-items: center;
+        width: 100%;
+        .titleInput {
+          padding: 0;
+          font-size: 14px;
+          background-color: transparent;
+          border: none;
+          box-shadow: none;
+        }
+        .loadingIcon {
+          flex-grow: 1;
+          margin-right: 12px;
+          visibility: visible;
+        }
+        .loadingIconHidden {
+          visibility: hidden;
+        }
+      }
+
+      .moreOpt {
+        display: flex;
+        justify-content: center;
+        align-items: center;
+        position: absolute;
+        top: 50%;
+        transform: translateY(-50%);
+        right: 0.75rem;
+        visibility: hidden;
+        width: 2rem;
+        height: 2rem;
+      }
+      &.active {
+        color: #2f69fe;
+        background: rgba(220, 230, 250);
+      }
+
+      &:hover {
+        color: #2f69fe;
+        background: rgb(220, 230, 250, 0.5);
+
+        .moreOpt {
+          visibility: visible;
+        }
+      }
+    }
+
+    .empty {
+      margin-top: 5rem;
+    }
+  }
+}
